# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voltpilot-EMS Portal API
  version: 0.1.0-mvp
  description: >
    OpenAPI contract for the portal backend (Spring Boot). BINDING contract
    (architecture section 20 item 8). Covers the MVP portal surface: auth-guarded
    tenancy, device claiming, telemetry and schedule views, and economic KPIs
    (architecture section 4).


    Every non-system endpoint requires a Keycloak-issued Bearer JWT (realm
    `voltpilot`). The `tenant_id` claim scopes all data via Postgres
    Row-Level-Security - a caller only ever sees its own tenant's rows; another
    tenant's resources return 404 (not found) rather than 403. The
    sites/devices/telemetry/prices/weather/schedule and device-claim endpoints
    are implemented in services/api; KPIs remain a stub pending a later
    increment.
  contact:
    name: Voltpilot-EMS
servers:
  - url: http://localhost:8090
    description: Local dev (docker compose / `mvn spring-boot:run`)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/devices:
    get:
      tags: [devices]
      summary: List devices for the caller's tenant.
      responses:
        "200":
          description: Devices (only the caller's tenant, via RLS).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/devices/{deviceId}:
    put:
      tags: [devices]
      summary: Update a device (type + label only; the ref is immutable).
      description: >
        Only `kind` and the optional customer-facing label `name` are editable.
        The `externalRef` is the device's identity (MQTT topics, registry
        gating, the printed sticker) and cannot be changed - a wrong ref is
        fixed by unclaiming and re-claiming.
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDeviceRequest" }
      responses:
        "200":
          description: Updated device.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "400": { description: Invalid kind or over-long label. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Device not found for this tenant. }
    delete:
      tags: [devices]
      summary: Unclaim (delete) a device incl. its telemetry + MQTT cleanup.
      description: >
        Removes the device row AND its telemetry, and clears the broker's
        retained `provision/{ref}/config` and schedule topics (best-effort,
        like the on-claim publish) so the physical device falls back to its
        watchdog default. The ref becomes claimable again; a sticker ID stays
        registered in the manufacturing registry.
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Device unclaimed; its telemetry deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Device not found for this tenant. }
  /api/v1/devices/claim:
    post:
      tags: [devices]
      summary: Claim an edge device into a site (onboarding = one insert).
      description: >
        Claims an unclaimed edge device (identified by `externalRef`) into one of
        the caller's sites. `externalRef` is canonicalized before use: trimmed,
        and sticker Geräte-IDs (prefix `VP-`, case-insensitive) uppercased, so
        the typed case never creates a second device. Sticker IDs additionally
        validate against the provisioned-device registry (manufactured IDs,
        maintained via the admin API): an unknown sticker ID yields 422 instead
        of creating a ghost device, and the claimed device inherits the
        registry's `kind`. Non-sticker refs stay ungated. The claim is idempotent
        within the tenant: re-claiming a ref the caller already owns returns 200
        with the existing device (at its original site) instead of a conflict.
        `externalRef` is globally unique, so a device claimed by ANOTHER tenant -
        which RLS hides - yields 409. The target site must belong to the caller's
        tenant, else 404.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceClaimRequest" }
      responses:
        "200":
          description: Already claimed by the caller's own tenant - the existing device.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "201":
          description: Device claimed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Target site not found for this tenant. }
        "409": { description: Device already claimed by another tenant. }
        "422": { description: Unknown sticker Geräte-ID (not in the provisioned-device registry). }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak-issued OIDC access token (realm `voltpilot`). Must carry a
        `tenant_id` claim; the API sets it as the Postgres `app.tenant_id`
        session variable to drive Row-Level-Security.

        Tenant switcher (Portal-Admins only): a token with the `platform-admin`
        realm role (which carries NO tenant_id) may select a tenant per request
        via the `X-Tenant-Id` header (UUID). The customer endpoints then serve
        that tenant's data through the same RLS scoping the customer gets. The
        header is IGNORED for every non-admin token, so it can never widen a
        customer's access; an admin without the header sees zero rows
        (default-deny).
  responses:
    Unauthorized:
      description: Missing, expired or invalid Bearer token.
...
  schemas:
...
    UpdateDeviceRequest:
      type: object
      description: Only kind + label are editable; the externalRef is immutable.
      properties:
        kind: { type: string, enum: [inverter, battery, meter], nullable: true, description: Unchanged when omitted. }
        name: { type: string, nullable: true, maxLength: 120, description: Customer-facing label; blank/null removes it. }
    DeviceClaimRequest:
      type: object
      required: [siteId, externalRef]
      properties:
        siteId: { type: string, format: uuid }
        externalRef: { type: string, description: Edge device identifier (globally unique). }
        kind: { type: string, description: "Device kind; defaults to 'inverter'.", example: inverter }
    Device:
      type: object
      properties:
        id: { type: string, format: uuid }
        siteId: { type: string, format: uuid }
        externalRef: { type: string }
        kind: { type: string, example: inverter }
        name:
          type: string
          nullable: true
          description: Optional customer-facing label (Bezeichnung); the immutable externalRef stays the identity.
        status: { type: string, example: claimed }
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Newest telemetry timestamp for the device; null until the first data
            arrives. The portal derives the onboarding state from it ("wartet auf
            erste Daten" -> online).
...